requirements.txt
streamlit>=1.38
python-docx

